+++ src/gamspy/_symbols/variable.py
@@ -270,7 +270,7 @@ class Variable(gt.Variable, operable.Operable, Symbol):
             self._stage = self._create_attr("stage")
 
             if records is not None:
-                self.setRecords(records)
+                self.setRecords(records, uels_on_axes=uels_on_axes)
             else:
                 self.container._run()